Base: implement separate drawables/selector for Holo Light Seekbars
This commit allows the framework to use separate drawables & xml selectors for seekbars on Holo & Holo.Light themes.
This is needed because the white drawables introduced in 5fdc52b are perfect on the dark theme, but almost invisible on the light one.
The darker holo light seekbars drawables for all resolutions are included in this commit.

Change-Id: I77e81ff624a0eaab5f91ec1f745f55d9d5dc9e4a